Fix issues when slicing a triangle geometry (#27)
* Fixing Issue with Triangles Above 0.0

There was an issue with a triangle geometry having a minimum coordinate in the slicing direction that is above 0.0.  This has been fied, but negative minimum coordinates is still an issue.  THis should be relooked at.

Signed-off-by: Michael Groeber<mike.groeber@bluequartz.net>

* Removal of methods that are not needed. Python bindings added

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>

* Clang-Format the code

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>

* Fixing a slicing inconsistency.

Simplified the slicing procedure and removed the dependency on GeometryMath. Added a check to make sure the slicing is consistent on all triangles.

Signed-off-by: Michael Groeber<mike.groeber@bluequartz.net>

* Use pragma once to protect headers

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>

* Addresses code review feedback

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>

Co-authored-by: Michael Groeber <michael.groeber@wpafb.af.mil>
Co-authored-by: COEIT\groeber.9 <groeber.9@osu.edu>